Begin actor syntax

+++ modules/Text/Character.d
@@ -1,3 +1,13 @@
 Character struct {
-  value: i16 // UTF16?
+  value: i32 // 22 bit
 }
+
+
+// A character is made up of one or more unicode scaler values
+
+// 🐥
+
+// grapheme cluster?
+
+// http://www.unicode.org/glossary/#unicode_scalar_value
+
